requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
billiard==4.2.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
kombu==5.3.3
numpy==1.26.1
pandas==2.1.2
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.9
Scrapy==2.6.3
Werkzeug==2.2.2
beautifulsoup4==4.11.1
crochet==2.0.0
elasticsearch==8.1.2
postmarker==1.0
pydash==5.1.0
pymongo==4.1.0
pymssql==2.2.8
pypyodbc==1.3.6
python_dateutil==2.8.2
random_user_agent==1.0.1
requests==2.32.3
retrying==1.3.4
scrapy-redis==0.7.3
scrapy-rotated-proxy==0.1.5
scrapy-user-agents==0.1.1
scrapy_selenium==0.0.7
user-agent==0.1.10
waitress==1.3.0
gevent==24.2.1
black==24.4.2
pre-commit==3.7.1
Flake8==7.0.0
pymssql==2.2.8
postmarker==1.0
pydash==5.1.0
celery_once==3.0.1
google-api-python-client==2.97.0
google-auth==2.22.0